requirements.txt
django==3.2
djangorestframework
django-cors-headers
mysqlclient
cloudinary
python-decouple
djangorestframework-simplejwt
#librerias para despliegue en la nube
gunicorn
psycopg2
dj-database-url
whitenoise